Move the interpolation step into a function
Improve the code organization by moving the interpolation step of the
timing recovery loop into a function. Review also the polyphase
implementation to support reusing the same function for all
interpolation methods. Instead of keeping a dedicated sample buffer for
the polyphase implementation, use the main vector of input samples.

Change also the order adopted for interpolator choices. Make the
polyphase implementation option "0" and the linear implementation as
option "1".