Support odd oversampling factors
The main issue with odd oversampling factors is dealing with the
zero-crossing (or early/late) interpolant used by the ZCTED, GTED, and
ELTED. For instance, if L=3, the zero-crossing interpolant would fall at
L/2=1.5, which is not a valid index.

This patch solves the problem by using ceil(L/2) instead and adding a
+-0.5 offset to the fractional timing interval mu when computing the
zero-crossing interpolant. It also adjusts the basepoint index whenever
the "mu +-0.5" value falls out of the [0, 1] range. Please refer to the
longer comments added in the code.

A side effect of this change is that the value of mu differs when
computing the main interpolant and the zero-crossing interpolant for odd
L. The reason is that the +-0.5 offset added to mu only applies to the
zero-crossing interpolant. Hence, when using the polyphase interpolator,
we can no longer use the same polyphase subfilter (dependent on mu) to
compute both interpolants. To facilitate, move the polyphase branch
selection logic to the interpolation function.