Make the final constellation plots optional
Make the constellation plots opened by the main.m script optional just
like the static plots from symbolTimingSync. Let them be enabled with
the debug_tl_static option. The main motivation is that often the main
script can be used to verify the MER changes due to some new design
choice on the symbol synchronizer loop, in which case the constellation
plots can be unnecessary.