Revert selection of polyphase branch using round
It seems more accurate to round the product "polyInterpFactor * mu"
instead of using the floor of this product for the polyphase branch
selection. However, the results don't indicate any significant
performance gain. It is a lot more important to have a high enough
polyInterpFactor (say, 128 instead of 32) than to use the round operator
instead of the floor counterpart.

On the flip side, if we use rounding, we would need the extra polyphase
subfilter, which makes the implementation less tractable. For instance,
this last subfilter has a shorter delay, so an implementation taking
care of subfilter delays would need to apply an exception for the last
filter. And so on. Besides, mu is by definition less than 1, so using
floor makes sure the selected branch is always within acceptable bounds.